requirements.txt
Django==4.2
psycopg2-binary==2.9.6
python-telegram-bot==20.8
djangorestframework==3.14.0
drf-yasg==1.21.5
python-dotenv==1.0.0
django-filter==23.2
asgiref==3.7.2
